// src/data/types/requests-types.ts
enum REQUESTS {
  WEB_AUTHENTICATION = "web-authentication",
  APP_AUTHENTICATION = "app-authentication",
  DEVICE_ACTIVATION = "all-devices-activation",
  GEO_RESTRICTION = "geo-restriction-policy",
  CANCELLATION_DTC = "cancellation-dtc",
  CANCELLATION_ITUNES = "cancellation-itunes",
  CANCELLATION_ROKU = "cancellation-roku",
  NO_REFUND_POLICY = "no-refund",
  REFUND_APPROVAL = "refund-approval",
  ADD_ON_STARZ = "add-on-starz",
  LANGUAGE_CHANGE = "language-change",
  PASSWORD_RESET = "password-reset",
  PAYMENT_INFORMATION_UPDATE = "payment-information-update",
  ACCOUNT_TRACKING = "account-tracking",
  ACCOUNT_DELETION = "account-deletion",
  GIFT_CARD_USE = "use-gift-card",
  GIFT_CARD_INFO = "gift-card-info",
  PROMO_CODE_REDEEM = "promo-code-redeem",
  PROMO_CODE_FAILURE = "promo-code-failure",
  PARTNER_LINKING = "partner-linking",
  DOWNLOAD_TO_GO = "download-to-go",
  CONTENT_REQUEST = "content-request",
  FEEDBACK = "feedback",
  MANAGE_PIN = "manage-pin",
  CANCELLATION_GOOGLE = "cancellation-google",
}

export default REQUESTS;