// src/data/types/categories-types.ts
enum CATEGORIES {
  ACCOUNT = "account",
  AUTHENTICATION = "authentication",
  CANCELLATION = "cancellation",
  DEVICE_ACTIVATION = "device-activation",
  GEO_RESTRICTION = "geo-restriction",
  GIFT_CARD = "gift-card",
  OTHER = "other",
  PROFIL = "profil",
  PROMO_CODE = "promo-code",
  PROBING = "probing",
  REFUND = "refund",
  SUBSCRIPTION = "subscription",
  SUGGESTION = "suggestion",
  TROUBLESHOOTING = "troubleshooting",
}
export default CATEGORIES;